requirements.txt
# Django Framework
Django==4.2.7

# PostgreSQL Database Adapter
psycopg2-binary==2.9.9

# Django REST Framework
djangorestframework==3.14.0

# JWT Authentication
djangorestframework-simplejwt==5.3.0

# CORS Headers for Frontend Integration
django-cors-headers==4.3.1

# HTTP Requests (for ETL and ClickHouse communication)
requests==2.31.0

# Python dotenv for environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0

# Development tools
ipython==8.17.2
